def backward_D_basic(self, netD, real, fake):         """Calculate GAN loss for the discriminator"""         # Real         pred_real = netD(real)         loss_D_real = self.criterionGAN(pred_real, True)         # Fake         pred_fake = netD(fake.detach())         loss_D_fake = self.criterionGAN(pred_fake, False)         # Combined loss (backward is handled in optimize_parameters to allow NaN/Inf skipping)         loss_D = (loss_D_real + loss_D_fake) * 0.5         return loss_D
    def backward_G(self):
        """Calculate the loss for generators G_A and G_B"""
        # GAN loss D_A(G_A(A))
        self.loss_G_A = self.criterionGAN(self.netD_A(self.fake_B), True)
        # GAN loss D_B(G_B(B))
        self.loss_G_B = self.criterionGAN(self.netD_B(self.fake_A), True)
        # Forward cycle loss || G_B(G_A(A)) - A||
        self.loss_cycle_A = self.criterionCycle(self.rec_A, self.real_A) * self.opt.lambda_A
        # Backward cycle loss || G_A(G_B(B)) - B||
        self.loss_cycle_B = self.criterionCycle(self.rec_B, self.real_B) * self.opt.lambda_B
        # Identity loss
        if self.opt.identity:
            self.loss_idt_A = self.criterionIdt(self.idt_A, self.real_B) * self.opt.lambda_B * self.opt.lambda_idt
            self.loss_idt_B = self.criterionIdt(self.idt_B, self.real_A) * self.opt.lambda_A * self.opt.lambda_idt
        else:
            self.loss_idt_A = 0
            self.loss_idt_B = 0
        # combined loss (backward is handled in optimize_parameters to allow NaN/Inf skipping)
        self.loss_G = self.loss_G_A + self.loss_G_B + self.loss_cycle_A + self.loss_cycle_B + self.loss_idt_A + self.loss_idt_B

    def optimize_parameters(self):
        """Calculate losses, gradients, and update network weights; called in every training iteration"""
        # forward
        self.forward()  # compute fake images and reconstruction images.

        # ------------------
        # Optimize Generators
        # ------------------
        self.set_requires_grad([self.netD_A, self.netD_B], False)
        self.optimizer_G.zero_grad(set_to_none=True)

        # Compute G losses (no backward inside backward_G)
        self.backward_G()

        # NaN / Inf check for generator loss BEFORE backward
        if torch.isnan(self.loss_G).any() or torch.isinf(self.loss_G).any():
            print("[WARN] NaN or Inf detected in Generator loss. Skipping G backward/step.")
            self.optimizer_G.zero_grad(set_to_none=True)
        else:
            self.loss_G.backward()
            self.optimizer_G.step()

        # ------------------
        # Optimize Discriminators
        # ------------------
        self.set_requires_grad([self.netD_A, self.netD_B], True)
        self.optimizer_D.zero_grad(set_to_none=True)

        # Compute D losses (no backward inside backward_D_basic)
        fake_B = self.fake_B_pool.query(self.fake_B)
        fake_A = self.fake_A_pool.query(self.fake_A)
        self.loss_D_A = self.backward_D_basic(self.netD_A, self.real_B, fake_B)
        self.loss_D_B = self.backward_D_basic(self.netD_B, self.real_A, fake_A)
        loss_D = self.loss_D_A + self.loss_D_B

        # NaN / Inf check for discriminator loss BEFORE backward
        if torch.isnan(loss_D).any() or torch.isinf(loss_D).any():
            print("[WARN] NaN or Inf detected in Discriminator loss. Skipping D backward/step.")
            self.optimizer_D.zero_grad(set_to_none=True)
        else:
            loss_D.backward()
            self.optimizer_D.step()

+++ test2.py
@@ -8,7 +8,7 @@ import cv2
 import numpy as np
 import torchvision.transforms as transforms
 
-netG_B = networks.define_G(1, 1, 128, 'resnet_9blocks').float()
+netG_B = networks.define_G(1, 1, 128, 'resnet_9blocks')
 
 state_dict = torch.load('/home/psdl/Workspace/pytorch-CycleGAN-and-pix2pix/checkpoints/maps_cyclegan/latest_net_G_A.pth', map_location='cuda:1')
